+++ src/main/java/nextstep/subway/controller/PathController.java
@@ -1,6 +1,6 @@
 package nextstep.subway.controller;
 
-import nextstep.subway.facade.SectionFacade;
+import nextstep.subway.facade.PathFacade;
 import nextstep.subway.service.response.PathResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,17 +10,17 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class PathController {
 
-    private final SectionFacade sectionFacade;
+    private final PathFacade pathFacade;
 
-    public PathController(SectionFacade sectionFacade) {
-        this.sectionFacade = sectionFacade;
+    public PathController(PathFacade pathFacade) {
+        this.pathFacade = pathFacade;
     }
 
     @GetMapping("/paths")
     public ResponseEntity<PathResponse> getPath(
         @RequestParam long source, @RequestParam long target) {
 
-        return ResponseEntity.ok(sectionFacade.getPath(source, target));
+        return ResponseEntity.ok(pathFacade.getPath(source, target));
     }
 }
